refactor: move alias packages
