+++ pkg/coredata/coredata.go
@@ -1,6 +1,6 @@
-package goa4web
+package coredata
 
 import common "github.com/arran4/goa4web/core/common"
 
-// CoreData exposes the common CoreData type at the top level package.
+// CoreData exposes the common CoreData type for external packages.
 type CoreData = common.CoreData
