+++ pkg/dbalias/dbalias.go
@@ -1,7 +1,8 @@
-package goa4web
+package dbalias
 
 import db "github.com/arran4/goa4web/internal/db"
 
+// These aliases expose selected internal db types.
 type (
 	DBTX                                                = db.DBTX
 	Queries                                             = db.Queries
@@ -13,4 +14,5 @@ type (
 	User                                                = db.User
 )
 
+// New returns a new Queries instance using the given database.
 func New(d db.DBTX) *Queries { return db.New(d) }
